Rename xengine ops module name
